requirements.txt
beautifulsoup4==4.8.0
bitly-api==0.3
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
idna==2.8
psycopg2==2.8.3
PyJWT==1.7.1
pypiwin32==223
PySocks==1.7.0
pytz==2019.2
pywin32==225
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
twilio==6.30.0
urllib3==1.25.3
win10toast==0.9
Pillow==6.1.0
pystray==0.14.4
infi.systray==0.1.11

